Move brokered services to a subfolder
Otherwise we have too many confusing names.
